drone-scheduling-engine/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
torch>=2.0.0
torch-geometric>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
websockets>=11.0
redis>=4.5.0
motor>=3.3.0
pymongo>=4.5.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.11.0
python-dotenv>=1.0.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
scikit-learn>=1.3.0
networkx>=3.1
